refactor project and add a readme file
