+++ activiti-cloud-identity-adapter/README.md
@@ -1,3 +1,10 @@
-# activiti-cloud-modeling
+# activiti-cloud-identity-adapter
 
-Activiti Cloud Modeling Service
+Activiti Cloud Identity Adapter Service
+
+This project aims to provide a common interface to contact any identity provider.
+
+To run this project you need to specify properties below:
+
+* ACT_KEYCLOAK_URL=https://your_keycloal_instance;
+* ACT_KEYCLOAK_REALM=alfresco
