+++ activiti-cloud-identity-adapter/src/main/java/org/activiti/cloud/identity/adapter/IdentityAdapterApplication.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.cloud.modeling;
+package org.activiti.cloud.identity.adapter;
 
 import org.activiti.cloud.identity.web.EnableIdentityManagementRestAPI;
 import org.springframework.boot.SpringApplication;
