+++ activiti-cloud-identity-adapter/src/main/resources/application.properties
@@ -4,4 +4,4 @@ spring.zipkin.base-url=http://zipkin:80/
 spring.zipkin.sender.type=web
 spring.sleuth.sampler.probability=1.0
 
-activiti.cloud.swagger.modeling-base-path=identity-adapter-service
+activiti.cloud.swagger.identity-adapter-base-path=identity-adapter-service
