+++ activiti-cloud-identity-adapter/src/test/java/org/activiti/cloud/identity/adapter/KeycloakManagementControllerIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.cloud.modeling;
+package org.activiti.cloud.identity.adapter;
 
 import static org.springframework.test.annotation.DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD;
